Fixed some warnings.
Added Jakob Riedles tiny-utf8 to credited libraries in README.md.

+++ src/Graphics/Renderer.cpp
@@ -307,6 +307,6 @@ namespace tml
     {
         return (m_vertexData.size() + vertexCount)  > s_maxVertexCount ||
                (m_indexData.size()  + indexCount)   > s_maxIndexCount  ||
-               (m_textures.size()   + textureCount) > m_maxTextureCount;
+               (m_textures.size()   + textureCount) > static_cast<uint32_t>(m_maxTextureCount);
     }
 }
